Change UsageEvent.Quantity from double to decimal
Decimal is more appropriate for monetary/metering quantities as it
avoids floating-point precision issues.

+++ src/AzureMarketplaceSandbox/Services/MeteringService.cs
@@ -132,7 +132,7 @@ public class MeteringService(MarketplaceDbContext db)
 public record UsageEventRequest
 {
     public Guid ResourceId { get; init; }
-    public double Quantity { get; init; }
+    public decimal Quantity { get; init; }
     public string Dimension { get; init; } = string.Empty;
     public DateTime EffectiveStartTime { get; init; }
     public string PlanId { get; init; } = string.Empty;
